+++ lib/QuantumToolboxCore/test/versioninfo_and_settings.jl
@@ -1,30 +1,30 @@
 @testitem "versioninfo and about" begin
     # test error handling
     io_wrong = IOBuffer()
-    push!(QuantumToolboxUtils._QT_LIBRARIES, Base)
-    @test_throws ArgumentError QuantumToolboxUtils.versioninfo(io_wrong)
-    deleteat!(QuantumToolboxUtils._QT_LIBRARIES, findall(x -> x == Base, QuantumToolboxUtils._QT_LIBRARIES))
+    push!(QuantumToolboxCore._QT_LIBRARIES, Base)
+    @test_throws ArgumentError QuantumToolboxCore.versioninfo(io_wrong)
+    deleteat!(QuantumToolboxCore._QT_LIBRARIES, findall(x -> x == Base, QuantumToolboxCore._QT_LIBRARIES))
 
     # versioninfo
     io_version = IOBuffer()
-    QuantumToolboxUtils.versioninfo(io_version)
+    QuantumToolboxCore.versioninfo(io_version)
     version_output = String(take!(io_version))
     @test occursin("QuantumToolbox.jl: Quantum Toolbox in Julia", version_output)
     @test occursin("Package information:", version_output)
-    @test occursin("QuantumToolboxUtils", version_output)
+    @test occursin("QuantumToolboxCore", version_output)
     @test occursin("System information:", version_output)
     @test occursin("Please cite QuantumToolbox.jl in your publication", version_output)
 
     # about
     io_about = IOBuffer()
-    QuantumToolboxUtils.about(io_about)
+    QuantumToolboxCore.about(io_about)
     about_output = String(take!(io_about))
     @test version_output == about_output
 end
 
 @testitem "Settings" begin
     io = IOBuffer()
-    show(io, QuantumToolboxUtils.settings)
+    show(io, QuantumToolboxCore.settings)
     out = String(take!(io))
     @test occursin("QuantumToolbox.jl Settings", out)
     @test occursin("tidyup_tol", out)