+++ test/runtests.jl
@@ -6,7 +6,7 @@ const testdir = dirname(@__FILE__)
 
 # Define the paths to the library
 const LIBRARY_NAME_AND_PATH = Dict(
-    "Utils" => ("QuantumToolboxUtils", joinpath(testdir, "..", "lib", "QuantumToolboxUtils")),
+    "Core" => ("QuantumToolboxCore", joinpath(testdir, "..", "lib", "QuantumToolboxCore")),
 )
 const LIBRARY_LIST = collect(keys(LIBRARY_NAME_AND_PATH))